+++ src/test/kotlin/com/github/pooryam92/vimcoach/features/tips/ui/notifications/TipNotificationServiceUiTest.kt
@@ -185,6 +185,9 @@ class TipNotificationServiceUiTest : BasePlatformTestCase() {
             return enabledCategories
         }
 
-        override fun setEnabledTipCategories(categories: List<String>) = Unit
+        override fun setEnabledTipCategories(
+            availableCategories: List<String>,
+            enabledCategories: List<String>
+        ) = Unit
     }
 }